v0.0.0-base.2 实现epub模式下载

+++ Api/models.py
@@ -3,5 +3,5 @@ from django.db import models
 
 class History(models.Model):
     file_name = models.CharField(max_length=255)
-    percent = models.IntegerField(default=0)
+    percent = models.FloatField(default=0)
     book_id = models.CharField(max_length=255)
